convert two tests
